# Sqlite3 parser perf improvements