… vendored/sqlite3-parser/src/parser/ast/mod.rs
… vendored/sqlite3-parser/src/parser/parse.y
… vendored/sqlite3-parser/third_party/lemon/lemon.c
… vendored/sqlite3-parser/third_party/lemon/lempar.rs